refactor(codegen): extract thunk and trampoline emitters into thunks.rs
Relocate the standalone synthesized-thunk/wrapper/trampoline emitters (~2.3k
lines) out of the llvm.rs god-module: the derived eq/hash thunk bodies, the
closure-env free thunk and call-context resolver, the tuple/collection in-place
thunk bodies, the owned-element and lambda-env drop thunks, the spawn-task
direct/closure wrappers and their name builders, and the actor terminate /
dispatch trampolines.

Pure relocation following the coro.rs carve: thunks imports the shared codegen
context and the clone/drop in-place machinery from crate::llvm, which calls back
via crate::thunks::*; suspend's calls into the relocated drop-thunk helpers are
re-qualified accordingly. Every emitted .ll is byte-identical (ll-corpus verify
clean, native and wasm32).
