+++ hew-codegen-rs/src/lib.rs
@@ -34,6 +34,7 @@ pub mod coro;
 pub mod llvm;
 pub(crate) mod runtime_abi;
 pub(crate) mod suspend;
+pub(crate) mod thunks;
 
 pub use llvm::{
     emit_module, emit_module_objects, validate_codegen_front, verify_pipeline, CodegenError,
